+++ tests/Factory/Fixtures/TestDiscoveryRecipeProvider.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of the puli/repository-manager package.
+ *
+ * (c) Bernhard Schussek <bschussek@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Puli\RepositoryManager\Tests\Factory\Fixtures;
+
+use Puli\RepositoryManager\Factory\BuildRecipe;
+use Puli\RepositoryManager\Factory\BuildRecipeProvider;
+use Puli\RepositoryManager\Factory\ProviderFactory;
+
+/**
+ * @since  1.0
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
+class TestDiscoveryRecipeProvider implements BuildRecipeProvider
+{
+    public function getRecipe($varName, $outputDir, $rootDir, array $options, ProviderFactory $providerFactory)
+    {
+        $recipe = new BuildRecipe();
+        $recipe->addImport(__NAMESPACE__.'\TestDiscovery');
+        $recipe->addVarDeclaration($varName, $varName.' = new TestDiscovery($repo);');
+
+        return $recipe;
+    }
+}