+++ tests/Factory/KeyValueStore/ArrayStoreRecipeProviderTest.php
@@ -1,0 +1,69 @@
+<?php
+
+/*
+ * This file is part of the puli/repository-manager package.
+ *
+ * (c) Bernhard Schussek <bschussek@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Puli\RepositoryManager\Tests\Factory\KeyValueStore;
+
+use Puli\RepositoryManager\Factory\KeyValueStore\ArrayStoreRecipeProvider;
+use Puli\RepositoryManager\Tests\Factory\AbstractRecipeProviderTest;
+
+/**
+ * @since  1.0
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
+class ArrayStoreRecipeProviderTest extends AbstractRecipeProviderTest
+{
+    /**
+     * @var ArrayStoreRecipeProvider
+     */
+    private $provider;
+
+    protected function setUp()
+    {
+        parent::setUp();
+
+        $this->provider = new ArrayStoreRecipeProvider();
+    }
+
+    public function testGetRecipe()
+    {
+        $recipe = $this->provider->getRecipe(
+            '$store',
+            $this->outputDir,
+            $this->rootDir,
+            array(),
+            $this->generatorFactory
+        );
+
+        $expected = <<<EOF
+\$store = new ArrayStore();
+EOF;
+
+        $this->assertCode($expected, $recipe);
+    }
+
+    public function testRunRecipe()
+    {
+        $recipe = $this->provider->getRecipe(
+            '$store',
+            $this->outputDir,
+            $this->rootDir,
+            array(),
+            $this->generatorFactory
+        );
+
+        $this->putCode($this->outputPath, $recipe);
+
+        require $this->outputPath;
+
+        $this->assertTrue(isset($store));
+        $this->assertInstanceOf('Webmozart\KeyValueStore\ArrayStore', $store);
+    }
+}